Add unit tests to GcpKmsClient.java.
To be able to unit-test this class, we need to add a function to inject a CloudKMS object. That function is package-private and can only be used in its own unit tests. It is not part of the public API.

Also, rename 'client' to 'cloudKms' because client is confusing. And add some missing annotations.

PiperOrigin-RevId: 476114844
Change-Id: Iccdee72cf6144173f1d7c8a41710793b03881700