Improve the refaster documentation somewhat.
Working on this I found ways to improve: We add a step in the script to go into a temp directory (which makes it easier to copy-paste). We also put the jars into a directory jars/ and copy the code into the directory refaster/. This makes it easier to see what's going on when one does it interactively. We finally apply the patch.

PiperOrigin-RevId: 517393925
Change-Id: Iac993ef08766dbfa05bba005d0c441a6c18a4e19