Document how Tink handles associatedData='' in AES SIV.
AES SIV allows the user to import a list of associated datas, but Tink only uses a single associated data in its interface. When the user inputs an empty associated data byte sequence, Tink will handle this as a list of one empty byte sequence, and not as an empty list.

PiperOrigin-RevId: 527491521
Change-Id: I6b316d35f82401179c205cc55bc58ba5252dda4e