Replace "sha256sum -c" with "sha256sum -c -" wherever it might be called on MacOS.
Trying to run the tests on MacOS Sequoia failed with errors such as:

++ sha256sum -c
usage: sha256sum [-bctwz] [files ...]

According to pages such as https://issues.apache.org/jira/browse/MWRAPPER-150 we have that the version of sha256sum which ships with macOS Sequoia has a bug and requires an additional "-" to explicitly specify we read from stdin.

Hence we add that.

PiperOrigin-RevId: 787061078
Change-Id: I50654431bc55c9a33105da1f43ae365e4a549f13