Add MODULE.bazel files for tink-java-gcpkms
With this change this project can be built with WORKSPACE (the default behavior set in `.bazelrc` files) or with Bzlmod (specifying `--enable_bzlmod`)). When `--enable_bzlmod` is set Bazel won't use the WORKSPACE files as it will be overwritten of the empty WORKSPACE.bzlmod files.
NOTE: This works only because this project now uses Bazel 7.
PiperOrigin-RevId: 824394656
Change-Id: I363c81735200c732d6e7d68084e685feb337449c